Refactor: change Agree comonents name
